extract transformers and common constants
