+++ commons.js
@@ -1,0 +1,16 @@
+// This HTML page contains all the necessary div
+const STATIC_URL = 'https://static-links-page.signalnerve.workers.dev'
+const AVATAR_URL = 'https://i.imgur.com/e1m6jaY.jpg'
+const NAME = 'Lynn Bao'
+const LINKS = [
+  { name: 'Personal', url: 'https://linruolynn.com' },
+  { name: 'GitHub', url: 'https://github.com/linnnruoo' },
+  { name: 'LinkedIn', url: 'https://www.linkedin.com/in/linruolynn/' },
+]
+
+module.exports = {
+  STATIC_URL,
+  AVATAR_URL,
+  NAME,
+  LINKS,
+}
